+++ packages/react/src/StyledProvider.tsx
@@ -7,25 +7,22 @@ import { platformSpecificSpaceUnits } from './utils';
 import { createGlobalStylesWeb } from './createGlobalStylesWeb';
 import { createGlobalStyles } from './createGlobalStyles';
 type Config = any;
+let colorModeSet = false;
 
 export const defaultConfig: { config: Config; colorMode: COLORMODES } = {
   config: {},
   colorMode: 'light',
 };
 
-// interface ConfigContextData {
-//   config: Config;
-//   setConfig: (config: Config) => void;
-// }
-
 const defaultContextData: Config = defaultConfig;
-
 const StyledContext = React.createContext<Config>(defaultContextData);
 
-// type IContext = {
-//   config: Config;
-//   colorMode?: COLORMODES;
-// };
+const setCurrentColorMode = (currentColorMode: string) => {
+  if (currentColorMode) {
+    set(currentColorMode === 'dark' ? 'dark' : 'light');
+    colorModeSet = true;
+  }
+};
 export const StyledProvider: React.FC<{
   config: Config;
   colorMode?: COLORMODES;
@@ -35,7 +32,6 @@ export const StyledProvider: React.FC<{
   const currentConfig = React.useMemo(() => {
     //TODO: Add this later
     return platformSpecificSpaceUnits(config, Platform.OS);
-    // return config;
   }, [config]);
 
   if (Platform.OS === 'web' && globalStyles) {
@@ -48,64 +44,40 @@ export const StyledProvider: React.FC<{
   }, [colorMode]);
 
   React.useEffect(() => {
-    if (currentColorMode) {
-      set(currentColorMode === 'dark' ? 'dark' : 'light');
-      onChange((currentColor: string) => {
-        // only for web
-        if (Platform.OS === 'web') {
-          if (currentColor === 'dark') {
-            document.documentElement.classList.remove(`gs-light`);
-          } else {
-            document.documentElement.classList.remove(`gs-dark`);
-          }
-          document.documentElement.classList.add(`gs-${currentColor}`);
-        }
-      });
-    }
-
-    if (Platform.OS === 'web') {
-      document.documentElement.classList.add(`gs-${get()}`);
-    }
-  }, [currentColorMode]);
-
-  React.useEffect(() => {
+    // Add gs class name
     if (Platform.OS === 'web') {
       document.documentElement.classList.add(`gs`);
     }
+
+    onChange((currentColor: string) => {
+      // only for web
+      if (Platform.OS === 'web') {
+        if (currentColor === 'dark') {
+          document.documentElement.classList.remove(`gs-light`);
+        } else {
+          document.documentElement.classList.remove(`gs-dark`);
+        }
+        document.documentElement.classList.add(`gs-${currentColor}`);
+      }
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  // Set colormode server side
+  React.useEffect(() => {
+    setCurrentColorMode(currentColorMode);
+  }, [currentColorMode]);
 
-  if (Platform.OS === 'web' && currentColorMode) {
-    set(currentColorMode === 'dark' ? 'dark' : 'light');
+  // Set colormode for the first time
+  if (!colorModeSet) {
+    setCurrentColorMode(currentColorMode);
   }
 
   const globalStyleMap =
     config.globalStyle && createGlobalStyles(config.globalStyle);
 
-  let contextValue;
-  if (Platform.OS === 'web') {
-    // This if statement technically breaks the rules of hooks, but is safe
-    // because the condition never changes after mounting.
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    contextValue = React.useMemo(() => {
-      return {
-        config: currentConfig,
-        globalStyle: globalStyleMap,
-      };
-    }, [currentConfig, globalStyleMap]);
-  } else {
-    // This if statement technically breaks the rules of hooks, but is safe
-    // because the condition never changes after mounting.
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    contextValue = React.useMemo(() => {
-      return {
-        config: currentConfig,
-        colorMode: currentColorMode,
-        globalStyle: globalStyleMap,
-      };
-    }, [currentConfig, currentColorMode, globalStyleMap]);
-  }
+  const contextValue = React.useMemo(() => {
+    return { config: currentConfig, globalStyle: globalStyleMap };
+  }, [currentConfig, globalStyleMap]);
 
   return (
     <StyledContext.Provider value={contextValue}>