+++ packages/react/src/utils.ts
@@ -75,12 +75,22 @@ export function resolveStringToken(
       return tokenValue;
     } else {
       if (tokenScaleMap[propName]) {
+        let modifiedTokenScale = token_scale;
         if (
-          config?.tokens[token_scale] &&
-          config?.tokens[token_scale].hasOwnProperty(splitCurrentToken[0])
+          token_scale === 'sizes' &&
+          !config?.tokens[token_scale]?.hasOwnProperty(splitCurrentToken[0])
+        ) {
+          modifiedTokenScale = 'space';
+        }
+
+        if (
+          config?.tokens[modifiedTokenScale] &&
+          config?.tokens[modifiedTokenScale].hasOwnProperty(
+            splitCurrentToken[0]
+          )
         ) {
           const tokenValue =
-            config?.tokens?.[token_scale]?.[splitCurrentToken[0]];
+            config?.tokens?.[modifiedTokenScale]?.[splitCurrentToken[0]];
           typeofResult = typeof tokenValue;
 
           if (typeof tokenValue !== 'undefined' && tokenValue !== null) {
@@ -112,6 +122,11 @@ export function resolveStringToken(
 export const getTokenFromConfig = (config: any, prop: any, value: any) => {
   const aliasTokenType = config.propertyTokenMap[prop];
 
+  let IsNegativeToken = false;
+  if (typeof value === 'string' && value.startsWith('-')) {
+    IsNegativeToken = true;
+    value = value.slice(1);
+  }
   // const tokenScale = config?.tokens?.[aliasTokenType];
   let token;
 
@@ -144,9 +159,15 @@ export const getTokenFromConfig = (config: any, prop: any, value: any) => {
     } else {
       token = value;
     }
-    // console.log(token, typeof token, prop, '******');
   }
 
+  if (IsNegativeToken) {
+    if (typeof token === 'number') {
+      token = -token;
+    } else if (typeof token === 'string') {
+      token = `-${token}`;
+    }
+  }
   return token;
 };
 
@@ -187,7 +208,6 @@ export function resolveTokensFromConfig(config: any, props: any) {
       value
     );
   });
-  // console.log('&&&&&', newProps);
 
   return newProps;
 }